The test assembly deploys NUnit.Engine.dll which depends on Mono.Cecil 0.9.6
Unfortunately the NUnit3TestAdapter NuGet package neither brings Cecil nor declares a dependency on it. But it is required for NUnit.Engine.dll to function.

The simplest is to just ship Cecil with the tests to keep the NUnit.Engine.dll happy.

Some test runners may be using their own nunit.engine.dll, so for them missing cecil wouldn't be a problem.