Don't export IViewTaggerProvider for MSBuildTextMateTagger.
The buffer tagger (ITaggerProvider) is sufficient and will properly tag the view. The buffer tagger will also properly work in scenarios that don't have a view, such as retrieving markers for the entire file for the vertical scrollbar.

A view tagger is primarily necessary when you need specific access to the view.

For this case it is redundant as all tags would appear twice, once from each tagger.