Import some internal code from MSBuild
Primarily intended for property function eval, but useful for other
places where we want behavioral compatibility.

The hash of the MSBuild repo at the time of import are explicitly noted in
the file headers, along with any changes that were made to the files after
importing them.

There are also some stubs to the limit amount of code imported and to try
to minimize modifications so the files are easier to update.