Rework context in MSBuildCompletionSource
Don't attach context to the session to pass it round, as
this can cause issues if the completion source is re-queried
in a single session.

Instead, pass around context on a new context mechanism provided
by XmlCompletionSource.

Documentation also depends on this context, so use a separate
documentation provider object that encapsulates it.