Schema inference now performs value kind inference
Instead of trying to infer value types from names in completion and quick info,
the schema inference should be responsible for this. We should not be inferring
values types for symbols from schemas - it's up to the schema to define them.

Special case completion for now so we still have completion for inferred
values types in the imported XSD schema. After updating the schema, this can
be removed.