Rework IDeprecatable into IVersionableSymbol
Symbols supporting this interface can now contain information
about the version that the symbol was deprecated or introduced.
We don't do anything with this info yet but having this
mechanism means we can now start annotating symbols with this
info in preparation for future use.

This info is stored on a SymbolVersionInfo object so it's easier
to expand in future without touching every implementation of
the interface. This also means that the increased memory cost is
paid only by the small number of versionable symbols that contain
version info.