Fix locating best ItemInfo from schemas
When other schemas add metadata to an existing item, they
essentially redefine it, but with only that metadata, and no type
or description. Metadata lookups merge the metadata from all
these defintions.

However, when locating the ItemInfo for the type/description, we
preferentially search for an ItemInfo that doesn't just have metadata.
This search was broken and is now fixed.