Fix non-Windows full solution build
The VS extension project has targets that call several tools
that only work on Windows. Disable those targets specifically
so that the full solution build works on non-Windows platforms.
It will not produce the VSIX but we can at least verify that
the VS extension project compiles after making changes in
shared projects it depends on.