Fix resolving System.Object from System.Runtime in task metadata builder
The way TaskMetadataBuilder loads info from assemblies is super hacky
but it'll do for now.

The original intent of using Roslyn was that it could load live metadata
for tasks that were open as source code in the host. That worked in VSMac
but has never worked in VSWin.

For the cases where we're loading tasks from assemblies we'd probably
be better off using System.Reflection.Metadata or something.

We also need to load the reference assemblies for the BCL in order to
find the docs XML, and bundle the Microsoft.Build.dll xml docs files.